backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
